fix: avoid sending response info to Webhook sender
The sender of webhooks doesn't have any real need to see our logging
messages. Instead, we should be sending those to our own log output. So
this commit introduces a bit of a restructuring around the standard
response format to deliver only the response code, and keep everything
else private.